# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend for XuXaDex Platform.",
        "title": "XuXaDex API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/tournament": {
            "post": {
                "description": "Create Tournaments handler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Create Tournaments",
                "parameters": [
                    {
                        "description": "Tournament create object",
                        "name": "tournament",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_tournament.TournamentCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_tournament.TournamentBaseEntity"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/tournament/dashboard": {
            "get": {
                "description": "Tournaments Dashboard handler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Tournaments Dashboard",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_tournament.TournamentDashboardDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/tournament/randomize": {
            "get": {
                "description": "Randomize Dates handler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Randomize Dates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/tournament/statuses": {
            "get": {
                "description": "Tournament Statuses handler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Tournament Statuses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_tournament.TournamentStatusEntity"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/tournament/{id}": {
            "get": {
                "description": "Tournament Get By ID handler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Tournament Get By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tournament ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_tournament.TournamentFullEntity"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/tournaments": {
            "get": {
                "description": "All Tournaments handler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "All Tournaments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "10",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "id:ASC",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "count\u003c100",
                        "name": "filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_tournament.TournamentBaseEntity"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_xuxadex_backend-mvp-main_internal_game.GameForMatchEntity": {
            "type": "object",
            "properties": {
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_xuxadex_backend-mvp-main_internal_player.PlayerEntity": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_user.UserBaseEntity"
                }
            }
        },
...
        "github_com_xuxadex_backend-mvp-main_internal_tournament.TournamentBaseEntity": {
            "type": "object",
            "properties": {
                "entrance_fee": {
                    "type": "number"
                },
                "free_slots": {
                    "type": "integer"
                },
                "game": {
                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_game.GameForMatchEntity"
                },
                "id": {
                    "type": "string"
                },
                "preview_url": {
                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_repository.NullString"
                },
                "start_timestamp": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "win_up_to": {
                    "type": "number"
                }
            }
        },
        "github_com_xuxadex_backend-mvp-main_internal_tournament.TournamentCreateDTO": {
            "type": "object",
            "properties": {
                "creator_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "entrance_fee": {
                    "type": "number"
                },
                "match_delay": {
                    "type": "integer"
                },
                "start_timestamp": {
                    "type": "string"
                },
                "team_size": {
                    "type": "integer"
                },
                "teams_count": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "github_com_xuxadex_backend-mvp-main_internal_tournament.TournamentDashboardDTO": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "ongoing": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_tournament.TournamentBaseEntity"
                    }
                },
                "start_soon": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_tournament.TournamentBaseEntity"
                    }
                },
                "upcoming": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_tournament.TournamentBaseEntity"
                    }
                }
            }
        },
        "github_com_xuxadex_backend-mvp-main_internal_tournament.TournamentFullEntity": {
            "type": "object",
            "properties": {
                "banner_url": {
                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_repository.NullString"
                },
                "entrance_fee": {
                    "type": "number"
                },
                "free_slots": {
                    "type": "integer"
                },
                "game": {
                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_game.GameForMatchEntity"
                },
                "id": {
                    "type": "string"
                },
                "players": {
                    "type": "object",
                    "properties": {
                        "count": {
                            "type": "integer"
                        },
                        "list": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_player.PlayerEntity"
                            }
                        }
                    }
                },
                "preview_url": {
                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_repository.NullString"
                },
                "prize_pool": {
                    "type": "number"
                },
                "start_timestamp": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "github_com_xuxadex_backend-mvp-main_internal_tournament.TournamentStatusEntity": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_xuxadex_backend-mvp-main_internal_user.UserBaseEntity": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
...
        "github_com_xuxadex_backend-mvp-main_pkg_repository.NullString": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        },
        "github_com_xuxadex_backend-mvp-main_pkg_responses.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    }
}